moved workflow into new nipoppy refactored structure
