+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "workflow/proc_pipe/tractoflow/tractoflow"]
-	path = workflow/proc_pipe/tractoflow/tractoflow
+	path = nipoppy/workflow/proc_pipe/tractoflow/tractoflow
 	url = git@github.com:scilus/tractoflow.git
