+++ nipoppy/workflow/catalog.py
