+++ nipoppy/workflow/dicom_org/sample_dicom_dir_func.py
