+++ nipoppy/workflow/dicom_org/shell_utils/download_from_loni.sh
