+++ nipoppy/workflow/dicom_org/shell_utils/make_squash.sh
