+++ nipoppy/workflow/dicom_org/shell_utils/unzip_loni_downloads.sh
