+++ nipoppy/workflow/proc_pipe/fmriprep/scripts/sample_fmriprep_slurm.sh
