+++ nipoppy/workflow/proc_pipe/mriqc/Dockerfile
