+++ nipoppy/workflow/tabular/RCFT_Copy_config.json
